requirements.txt
google-adk==1.14.1
ai-boutique-assit-mcp==1.0.4